Refactor Methods
