+++ GRBL-Plotter/_misc/tools.csv
@@ -1,0 +1,12 @@
+1;FFFF00;gelb (44);0;0;0;2;1000;-1;100;1000;100
+2;FAD780;apricot (26);0;-15;0;2;1000;-1;100;1000;100
+3;FFA500;orange (54);0;-30;0;2;1000;-1;100;1000;100
+4;4CBED9;hellblau (57);0;-45;0;2;1000;-1;100;1000;100
+5;14B43C;hellgrün (43);0;-60;0;2;1000;-1;100;1000;100
+6;FF0000;rot (40);0;-75;0;2;1000;-1;100;1000;100
+7;D90385;pink (56);0;-90;0;2;1000;-1;100;1000;100
+8;8C5F96;flieder (59);0;-105;0;2;1000;-1;100;1000;100
+9;008000;grün (36);0;-120;0;2;1000;-1;100;1000;100
+10;0000FF;blau (41);0;-135;0;2;1000;-1;100;1000;100
+11;5F311E;braun (45);0;-150;0;2;1000;-1;100;1000;100
+12;000000;schwarz (46);0;-165;0;2;1000;-1;100;1000;100
